requirements.txt
streamlit
pandas
dnspython
email-validator